requirements.txt
apache-beam==2.61.0
apache-flink==2.1.0
apache-flink-libraries==2.1.0
attrs==25.4.0
avro==1.12.0
certifi==2025.10.5
charset-normalizer==3.4.3
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.8.0
docopt==0.6.2
fastavro==1.12.1
fasteners==0.20
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.2.15
numpy==2.1.3
objsize==0.7.1
orjson==3.11.3
packaging==25.0
pandas==2.3.3
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.7
pydot==1.4.2
PyJWT==2.10.1
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zstandard==0.25.0
